Improve the separation between the Connect and MM2 operators (#9955)
