+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/AbstractConnectOperator.java
@@ -6,14 +6,12 @@ package io.strimzi.operator.cluster.operator.assembly;
 
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.DefaultKubernetesResourceList;
-import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.LocalObjectReference;
 import io.fabric8.kubernetes.api.model.ServiceAccount;
 import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicy;
 import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBinding;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import io.micrometer.core.instrument.Timer;
 import io.strimzi.api.kafka.model.common.Condition;
 import io.strimzi.api.kafka.model.common.ConnectorState;
 import io.strimzi.api.kafka.model.connect.AbstractKafkaConnectSpec;
@@ -21,7 +19,6 @@ import io.strimzi.api.kafka.model.connect.KafkaConnectStatus;
 import io.strimzi.api.kafka.model.connector.AutoRestartStatus;
 import io.strimzi.api.kafka.model.connector.AutoRestartStatusBuilder;
 import io.strimzi.api.kafka.model.connector.KafkaConnector;
-import io.strimzi.api.kafka.model.connector.KafkaConnectorBuilder;
 import io.strimzi.api.kafka.model.connector.KafkaConnectorList;
 import io.strimzi.api.kafka.model.connector.KafkaConnectorSpec;
 import io.strimzi.api.kafka.model.connector.KafkaConnectorStatus;
@@ -33,7 +30,6 @@ import io.strimzi.operator.cluster.model.ImagePullPolicy;
 import io.strimzi.operator.cluster.model.KafkaConnectCluster;
 import io.strimzi.operator.cluster.model.KafkaConnectorConfiguration;
 import io.strimzi.operator.cluster.model.KafkaVersion;
-import io.strimzi.operator.cluster.model.NoSuchResourceException;
 import io.strimzi.operator.cluster.model.PodSetUtils;
 import io.strimzi.operator.cluster.model.RestartReason;
 import io.strimzi.operator.cluster.model.RestartReasons;
@@ -54,7 +50,6 @@ import io.strimzi.operator.common.Annotations;
 import io.strimzi.operator.common.BackOff;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
-import io.strimzi.operator.common.model.InvalidResourceException;
 import io.strimzi.operator.common.model.Labels;
 import io.strimzi.operator.common.model.StatusDiff;
 import io.strimzi.operator.common.model.StatusUtils;
@@ -66,18 +61,14 @@ import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.TreeMap;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static io.strimzi.operator.common.Annotations.ANNO_STRIMZI_IO_RESTART;
-import static io.strimzi.operator.common.Annotations.ANNO_STRIMZI_IO_RESTART_TASK;
 import static java.util.Collections.emptyMap;
 
 /**
@@ -89,7 +80,7 @@ import static java.util.Collections.emptyMap;
  * @param <P>   Custom Resource Spec type
  * @param <S>   Custom Resource Status type
  */
-@SuppressWarnings({"checkstyle:ClassFanOutComplexity", "checkstyle:CyclomaticComplexity"})
+@SuppressWarnings({"checkstyle:ClassFanOutComplexity"})
 public abstract class AbstractConnectOperator<C extends KubernetesClient, T extends CustomResource<P, S>,
         L extends DefaultKubernetesResourceList<T>, P extends AbstractKafkaConnectSpec, S extends KafkaConnectStatus>
         extends AbstractOperator<T, P, S, CrdOperator<C, T, L>> {
@@ -173,44 +164,6 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
     }
 
     /**
-     * Checks whether the use of KafkaConnector resources is enabled for this cluster or not
-     *
-     * @param connect   The Connect custom resource
-     *
-     * @return  True if the connector operator is enabled and KafkaConnector resources should be used. False otherwise.
-     */
-    public static boolean isUseResources(HasMetadata connect) {
-        return Annotations.booleanAnnotation(connect, Annotations.STRIMZI_IO_USE_CONNECTOR_RESOURCES, false);
-    }
-
-    /**
-     * Creates an exception indicating that the KafkaConnect resource which would match given KafkaConnector resource
-     * does not exist.
-     *
-     * @param connectNamespace  Namespace where the KafkaConnect resource is missing
-     * @param connectName       Name of the missing KafkaConnect resource
-     *
-     * @return  NoSuchResourceException indicating that the Connect cluster is missing
-     */
-    static NoSuchResourceException noConnectCluster(String connectNamespace, String connectName) {
-        return new NoSuchResourceException(
-                "KafkaConnect resource '" + connectName + "' identified by label '" + Labels.STRIMZI_CLUSTER_LABEL + "' does not exist in namespace " + connectNamespace + ".");
-    }
-
-    /**
-     * Creates an exception indicating that the connector cannot run because its Connect cluster has 0 replicas.
-     *
-     * @param connectNamespace  Namespace where the KafkaConnect resource exists
-     * @param connectName       Name of the KafkaConnect resource
-     *
-     * @return  NoSuchResourceException indicating that the Connect cluster is missing
-     */
-    static RuntimeException zeroReplicas(String connectNamespace, String connectName) {
-        return new RuntimeException(
-                "Kafka Connect cluster '" + connectName + "' in namespace " + connectNamespace + " has 0 replicas.");
-    }
-
-    /**
      * Reconciles the ServiceAccount for the Connect cluster.
      *
      * @param reconciliation       The reconciliation
@@ -286,89 +239,6 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
                 .compose(metricsAndLoggingCm -> Future.succeededFuture(kafkaConnectCluster.generateMetricsAndLogConfigMap(metricsAndLoggingCm)));
     }
 
-    protected KafkaConnectApi getKafkaConnectApi() {
-        return connectClientProvider.apply(vertx);
-    }
-
-    /*test*/ Future<Void> reconcileConnectorAndHandleResult(Reconciliation reconciliation, String host, KafkaConnectApi apiClient,
-                                             boolean useResources, String connectorName, KafkaConnector connector) {
-        Promise<Void> reconciliationResult = Promise.promise();
-
-        metrics().connectorsReconciliationsCounter(reconciliation.namespace()).increment();
-        Timer.Sample connectorsReconciliationsTimerSample = Timer.start(metrics().metricsProvider().meterRegistry());
-
-        if (connector != null && Annotations.isReconciliationPausedWithAnnotation(connector)) {
-            return maybeUpdateConnectorStatus(reconciliation, connector, null, null).compose(
-                i -> {
-                    connectorsReconciliationsTimerSample.stop(metrics().connectorsReconciliationsTimer(reconciliation.namespace()));
-                    metrics().connectorsSuccessfulReconciliationsCounter(reconciliation.namespace()).increment();
-                    return Future.succeededFuture();
-                }
-            );
-        }
-
-        reconcileConnector(reconciliation, host, apiClient, useResources, connectorName, connector)
-                .onComplete(result -> {
-                    if (result.succeeded() && result.result() == null)  {
-                        // The reconciliation succeeded, but there is no status to be set => we complete the reconciliation and return
-                        // This normally means that the connector was deleted and there is no status to be set
-                        metrics().connectorsSuccessfulReconciliationsCounter(reconciliation.namespace()).increment();
-                        connectorsReconciliationsTimerSample.stop(metrics().connectorsReconciliationsTimer(reconciliation.namespace()));
-                        reconciliationResult.complete();
-                    } else if (result.failed() && connector == null) {
-                        // The reconciliation failed on connector deletion, so there is nowhere for status to be set => we complete the reconciliation and return
-                        LOGGER.warnCr(reconciliation, "Error reconciling connector {}", connectorName, result.cause());
-                        metrics().connectorsFailedReconciliationsCounter(reconciliation.namespace()).increment();
-                        connectorsReconciliationsTimerSample.stop(metrics().connectorsReconciliationsTimer(reconciliation.namespace()));
-
-                        // We suppress the error to not fail Connect reconciliation just because of a failing connector
-                        reconciliationResult.complete();
-                    } else {
-                        maybeUpdateConnectorStatus(reconciliation, connector, result.result(), result.cause())
-                                .onComplete(statusResult -> {
-                                    connectorsReconciliationsTimerSample.stop(metrics().connectorsReconciliationsTimer(reconciliation.namespace()));
-
-                                    if (result.succeeded() && statusResult.succeeded()) {
-                                        metrics().connectorsSuccessfulReconciliationsCounter(reconciliation.namespace()).increment();
-                                        reconciliationResult.complete();
-                                    } else {
-                                        // Reconciliation failed if either reconciliation or status update failed
-                                        metrics().connectorsFailedReconciliationsCounter(reconciliation.namespace()).increment();
-
-                                        // We suppress the error to not fail Connect reconciliation just because of a failing connector
-                                        reconciliationResult.complete();
-                                    }
-                                });
-                    }
-                });
-
-        return reconciliationResult.future();
-    }
-
-    private Future<ConnectorStatusAndConditions> reconcileConnector(Reconciliation reconciliation, String host, KafkaConnectApi apiClient,
-                                             boolean useResources, String connectorName, KafkaConnector connector) {
-        if (connector == null) {
-            if (useResources) {
-                LOGGER.infoCr(reconciliation, "deleting connector: {}", connectorName);
-                return apiClient.delete(reconciliation, host, port, connectorName).mapEmpty();
-            } else {
-                return Future.succeededFuture();
-            }
-        } else {
-            LOGGER.infoCr(reconciliation, "creating/updating connector: {}", connectorName);
-
-            if (connector.getSpec() == null) {
-                return Future.failedFuture(new InvalidResourceException("spec property is required"));
-            }
-
-            if (!useResources) {
-                return Future.failedFuture(new NoSuchResourceException(reconciliation.kind() + " " + reconciliation.name() + " is not configured with annotation " + Annotations.STRIMZI_IO_USE_CONNECTOR_RESOURCES));
-            } else {
-                return maybeCreateOrUpdateConnector(reconciliation, host, apiClient, connectorName, connector.getSpec(), connector);
-            }
-        }
-    }
-
     /**
      * Try to get the current connector config. If the connector does not exist, or its config differs from the
      * {@code connectorSpec}'s, then call
@@ -441,7 +311,7 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
         return !desiredConfig.equals(actualConfig);
     }
 
-    protected Future<Map<String, Object>> createOrUpdateConnector(Reconciliation reconciliation, String host, KafkaConnectApi apiClient,
+    private Future<Map<String, Object>> createOrUpdateConnector(Reconciliation reconciliation, String host, KafkaConnectApi apiClient,
                                                                   String connectorName, KafkaConnectorSpec connectorSpec, KafkaConnectorConfiguration desiredConfig) {
         return apiClient.createOrUpdatePutRequest(reconciliation, host, port, connectorName, asJson(connectorSpec, desiredConfig))
             .compose(ignored -> apiClient.statusWithBackOff(reconciliation, new BackOff(200L, 2, 10), host, port,
@@ -633,7 +503,7 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
      *
      * @return  Number of minutes after which the next restart should happen
      */
-    private int nextAutoRestartBackOffIntervalInMinutes(int restartCount)    {
+    private static int nextAutoRestartBackOffIntervalInMinutes(int restartCount)    {
         return Math.min(restartCount * restartCount + restartCount, 60);
     }
 
@@ -646,7 +516,7 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
      *
      * @return  True if the previous auto-restart status of the connector should be reset to 0. False otherwise.
      */
-    /* test */ boolean shouldResetAutoRestartStatus(AutoRestartStatus autoRestartStatus) {
+    /* test */ static boolean shouldResetAutoRestartStatus(AutoRestartStatus autoRestartStatus) {
         if (autoRestartStatus != null
                 && autoRestartStatus.getLastRestartTimestamp() != null
                 && autoRestartStatus.getCount() > 0) {
@@ -705,182 +575,86 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
     }
 
     /**
-     * Whether the provided resource instance is a KafkaConnector and has the strimzi.io/restart annotation
+     * Creates the ClusterRoleBinding required for the init container used for client rack-awareness.
+     * The init-container needs to be able to read the labels from the node it is running on to be able
+     * to determine the `client.rack` option.
      *
-     * @param resource resource instance to check
-     * @param connectorName connectorName name of the connector to check
-     * @return true if the provided resource instance has the strimzi.io/restart annotation; false otherwise
-     */
-    @SuppressWarnings({ "rawtypes" })
-    protected boolean hasRestartAnnotation(CustomResource resource, String connectorName) {
-        return Annotations.booleanAnnotation(resource, ANNO_STRIMZI_IO_RESTART, false);
-    }
-
-    /**
-     * Return the ID of the connector task to be restarted if the provided KafkaConnector resource instance has the strimzi.io/restart-task annotation
+     * @param reconciliation The reconciliation
+     * @param crbName ClusterRoleBinding name
+     * @param crb ClusterRoleBinding
      *
-     * @param resource resource instance to check
-     * @param connectorName KafkaConnector resource instance to check
-     * @return the ID of the task to be restarted if the provided KafkaConnector resource instance has the strimzi.io/restart-task annotation or -1 otherwise.
+     * @return Future for tracking the asynchronous result of the ClusterRoleBinding reconciliation
      */
-    @SuppressWarnings({ "rawtypes" })
-    protected int getRestartTaskAnnotationTaskID(CustomResource resource, String connectorName) {
-        return Annotations.intAnnotation(resource, ANNO_STRIMZI_IO_RESTART_TASK, -1);
+    protected Future<ReconcileResult<ClusterRoleBinding>> connectInitClusterRoleBinding(Reconciliation reconciliation, String crbName, ClusterRoleBinding crb) {
+        return ReconcilerUtils.withIgnoreRbacError(reconciliation, clusterRoleBindingOperations.reconcile(reconciliation, crbName, crb), crb);
     }
 
-    /**
-     * Patches the KafkaConnector CR to remove the strimzi.io/restart annotation, as
-     * the restart action specified by the user has been completed.
-     */
-    @SuppressWarnings({ "rawtypes" })
-    protected Future<Void> removeRestartAnnotation(Reconciliation reconciliation, CustomResource resource) {
-        return removeAnnotation(reconciliation, (KafkaConnector) resource, ANNO_STRIMZI_IO_RESTART);
-    }
+    // Abstract methods for working with restart annotations. These methods are implemented in the KafkaConnectAssemblyOperator and KafkaMirrorMaker2AssemblyOperator
 
     /**
-     * Patches the KafkaConnector CR to remove the strimzi.io/restart-task annotation, as
-     * the restart action specified by the user has been completed.
-     */
-    @SuppressWarnings({ "rawtypes" })
-    protected Future<Void> removeRestartTaskAnnotation(Reconciliation reconciliation, CustomResource resource) {
-        return removeAnnotation(reconciliation, (KafkaConnector) resource, ANNO_STRIMZI_IO_RESTART_TASK);
-    }
-
-    /**
-     * Patches the KafkaConnector CR to remove the supplied annotation.
-     */
-    private Future<Void> removeAnnotation(Reconciliation reconciliation, KafkaConnector resource, String annotationKey) {
-        LOGGER.debugCr(reconciliation, "Removing annotation {}", annotationKey);
-        KafkaConnector patchedKafkaConnector = new KafkaConnectorBuilder(resource)
-            .editMetadata()
-            .removeFromAnnotations(annotationKey)
-            .endMetadata()
-            .build();
-        return connectorOperator.patchAsync(reconciliation, patchedKafkaConnector)
-            .compose(ignored -> Future.succeededFuture());
-    }
-
-    /**
-     * Update the status of the Kafka Connector custom resource
+     * Checks whether the provided resource instance (a KafkaConnector or KafkaMirrorMaker2) has the annotation requesting a restart.
      *
-     * @param reconciliation        Reconciliation marker
-     * @param error                 Throwable indicating if any errors occurred during the reconciliation
-     * @param kafkaConnector2       Latest version of the KafkaConnector resource where the status should be updated
-     * @param connectorOperations   The KafkaConnector operations for updating the status
+     * @param resource          Resource instance to check
+     * @param connectorName     Connector name of the connector to check
+     *
+     * @return True if the provided resource has the restart annotation. False otherwise.
      */
-    public static void updateStatus(Reconciliation reconciliation, Throwable error, KafkaConnector kafkaConnector2, CrdOperator<?, KafkaConnector, ?> connectorOperations) {
-        KafkaConnectorStatus status = new KafkaConnectorStatus();
-        StatusUtils.setStatusConditionAndObservedGeneration(kafkaConnector2, status, error);
-        StatusDiff diff = new StatusDiff(kafkaConnector2.getStatus(), status);
-        if (!diff.isEmpty()) {
-            KafkaConnector copy = new KafkaConnectorBuilder(kafkaConnector2).build();
-            copy.setStatus(status);
-            connectorOperations.updateStatusAsync(reconciliation, copy);
-        }
+    @SuppressWarnings({ "rawtypes" })
+    abstract boolean hasRestartAnnotation(CustomResource resource, String connectorName);
+
+    /**
+     * Returns the ID of the connector task to be restarted from the (a KafkaConnector or KafkaMirrorMaker2) custom resource.
+     *
+     * @param resource          Resource instance to check
+     * @param connectorName     Connector name to check
+     *
+     * @return  The ID of the task to be restarted based on the annotation or -1 otherwise.
+     */
+    @SuppressWarnings({ "rawtypes" })
+    abstract int getRestartTaskAnnotationTaskID(CustomResource resource, String connectorName);
+
+    /**
+     * Patches the custom resource to remove the restart annotation
+     *
+     * @param reconciliation    Reconciliation marker
+     * @param resource          Custom resource from which the annotation should be removed
+     *
+     * @return  Future that indicates the operation completion
+     */
+    @SuppressWarnings({ "rawtypes" })
+    abstract Future<Void> removeRestartAnnotation(Reconciliation reconciliation, CustomResource resource);
+
+    /**
+     * Patches the custom resource to remove the restart task annotation
+     *
+     * @param reconciliation    Reconciliation marker
+     * @param resource          Custom resource from which the annotation should be removed
+     *
+     * @return  Future that indicates the operation completion
+     */
+    @SuppressWarnings({ "rawtypes" })
+    abstract Future<Void> removeRestartTaskAnnotation(Reconciliation reconciliation, CustomResource resource);
+
+    // Static utility methods and classes
+
+    private static Function<Map<String, Object>, Future<ConnectorStatusAndConditions>> createConnectorStatusAndConditions() {
+        return statusResult -> Future.succeededFuture(new ConnectorStatusAndConditions(statusResult, List.of(), List.of(), null));
     }
 
-    static protected class ConnectorStatusAndConditions {
-        Map<String, Object> statusResult;
-        List<String> topics;
-        List<Condition> conditions;
-        AutoRestartStatus autoRestart;
-
-        ConnectorStatusAndConditions(Map<String, Object> statusResult, List<String> topics, List<Condition> conditions, AutoRestartStatus autoRestart) {
-            this.statusResult = statusResult;
-            this.topics = topics;
-            this.conditions = conditions;
-            this.autoRestart = autoRestart;
-        }
-
-        ConnectorStatusAndConditions(Map<String, Object> statusResult, List<Condition> conditions) {
-            this(statusResult, Collections.emptyList(), conditions, null);
-        }
-
-        ConnectorStatusAndConditions(Map<String, Object> statusResult) {
-            this(statusResult, Collections.emptyList(), Collections.emptyList(), null);
-        }
+    private static Function<Map<String, Object>, Future<ConnectorStatusAndConditions>> createConnectorStatusAndConditions(List<Condition> conditions) {
+        return statusResult -> Future.succeededFuture(new ConnectorStatusAndConditions(statusResult, List.of(), conditions, null));
     }
 
-    Function<Map<String, Object>, Future<ConnectorStatusAndConditions>> createConnectorStatusAndConditions() {
-        return statusResult -> Future.succeededFuture(new ConnectorStatusAndConditions(statusResult));
-    }
-
-    Function<Map<String, Object>, Future<ConnectorStatusAndConditions>> createConnectorStatusAndConditions(List<Condition> conditions) {
-        return statusResult -> Future.succeededFuture(new ConnectorStatusAndConditions(statusResult, conditions));
-    }
-
-    Function<List<String>, Future<ConnectorStatusAndConditions>> updateConnectorStatusAndConditions(ConnectorStatusAndConditions status) {
+    private static Function<List<String>, Future<ConnectorStatusAndConditions>> updateConnectorStatusAndConditions(ConnectorStatusAndConditions status) {
         return topics -> Future.succeededFuture(new ConnectorStatusAndConditions(status.statusResult, topics, status.conditions, status.autoRestart));
     }
 
-    /**
-     * Validates the KafkaConnector resource and returns any warning conditions find during the validation
-     *
-     * @param reconciliation    Reconciliation marker
-     * @param resource          KafkaConnector resource which should be validated
-     *
-     * @return  Set with the warning conditions. Empty set if no conditions were found.
-     */
-    public Set<Condition> validate(Reconciliation reconciliation, KafkaConnector resource) {
-        if (resource != null) {
-            return StatusUtils.validate(reconciliation, resource);
-        }
-
-        return Collections.emptySet();
-    }
-
-    Future<Void> maybeUpdateConnectorStatus(Reconciliation reconciliation, KafkaConnector connector, ConnectorStatusAndConditions connectorStatus, Throwable error) {
-        KafkaConnectorStatus status = new KafkaConnectorStatus();
-        if (error != null) {
-            LOGGER.warnCr(reconciliation, "Error reconciling connector {}", connector.getMetadata().getName(), error);
-        }
-
-        Map<String, Object> statusResult = null;
-        List<String> topics = null;
-        List<Condition> conditions = new ArrayList<>();
-        AutoRestartStatus autoRestart = null;
-
-        Future<Void> connectorReadiness = Future.succeededFuture();
-
-        if (connectorStatus != null) {
-            statusResult = connectorStatus.statusResult;
-            JsonObject statusResultJson = new JsonObject(statusResult);
-            List<String> failedTaskIds;
-            if (connectorHasFailed(statusResultJson)) {
-                connectorReadiness = Future.failedFuture(new Throwable("Connector has failed, see connectorStatus for more details."));
-            } else if (!(failedTaskIds = failedTaskIds(statusResultJson)).isEmpty()) {
-                connectorReadiness = Future.failedFuture(new Throwable(String.format("The following tasks have failed: %s, see connectorStatus for more details.", String.join(", ", failedTaskIds))));
-            }
-            topics = connectorStatus.topics.stream().sorted().collect(Collectors.toList());
-            conditions.addAll(connectorStatus.conditions);
-            autoRestart = connectorStatus.autoRestart;
-        }
-
-        Set<Condition> unknownAndDeprecatedConditions = validate(reconciliation, connector);
-        conditions.addAll(unknownAndDeprecatedConditions);
-
-        if (!Annotations.isReconciliationPausedWithAnnotation(connector)) {
-            StatusUtils.setStatusConditionAndObservedGeneration(connector, status, error != null ? error : connectorReadiness.cause());
-            status.setConnectorStatus(statusResult);
-            status.setTasksMax(getActualTaskCount(connector, statusResult));
-            status.setTopics(topics);
-            status.setAutoRestart(autoRestart);
-        } else {
-            status.setObservedGeneration(connector.getStatus() != null ? connector.getStatus().getObservedGeneration() : 0);
-            conditions.add(StatusUtils.getPausedCondition());
-        }
-        status.addConditions(conditions);
-
-        return maybeUpdateStatusCommon(connectorOperator, connector, reconciliation, status,
-            (connector1, status1) -> new KafkaConnectorBuilder(connector1).withStatus(status1).build());
-    }
-
-    private boolean connectorHasFailed(JsonObject statusResult) {
+    protected static boolean connectorHasFailed(JsonObject statusResult) {
         JsonObject connectorStatus = statusResult.getJsonObject("connector");
         return connectorStatus != null && "FAILED".equals(connectorStatus.getString("state"));
     }
 
-    private List<String> failedTaskIds(JsonObject statusResult) {
+    protected static List<String> failedTaskIds(JsonObject statusResult) {
         JsonArray tasks = Optional.ofNullable(statusResult.getJsonArray("tasks")).orElse(new JsonArray());
         List<String> failedTasks = new ArrayList<>();
         for (Object task : tasks) {
@@ -891,30 +665,7 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
         return failedTasks;
     }
 
-    /**
-     * The tasksMax are mirrored in the KafkaConnector status section where they are used by the scale subresource.
-     * However, .spec.tasksMax is not always set and has no default value in Strimzi (only in Kafka Connect). So when
-     * it is not set, we try to count the tasks from the status. And if these are missing as well, we just set it to 0.
-     *
-     * @param connector         The KafkaConnector instance of the reconciled connector
-     * @param statusResult      The status from the Connect REST API
-     * @return                  Number of tasks which should be set in the status
-     */
-    @SuppressWarnings({ "rawtypes" })
-    protected int getActualTaskCount(KafkaConnector connector, Map<String, Object> statusResult)  {
-        if (connector.getSpec() != null
-                && connector.getSpec().getTasksMax() != null)  {
-            return connector.getSpec().getTasksMax();
-        } else if (statusResult != null
-                && statusResult.containsKey("tasks")
-                && statusResult.get("tasks") instanceof List) {
-            return ((List) statusResult.get("tasks")).size();
-        } else {
-            return 0;
-        }
-    }
-
-    protected JsonObject asJson(KafkaConnectorSpec spec, KafkaConnectorConfiguration desiredConfig) {
+    private static JsonObject asJson(KafkaConnectorSpec spec, KafkaConnectorConfiguration desiredConfig) {
         JsonObject connectorConfigJson = new JsonObject();
 
         for (Map.Entry<String, String> cf : desiredConfig.asOrderedProperties().asMap().entrySet()) {
@@ -929,7 +680,7 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
     }
 
     @SuppressWarnings({ "rawtypes" })
-    protected Future<KafkaConnectorStatus> getPreviousKafkaConnectorStatus(Reconciliation reconciliation, CrdOperator<KubernetesClient, KafkaConnector, KafkaConnectorList> resourceOperator,
+    private static Future<KafkaConnectorStatus> getPreviousKafkaConnectorStatus(Reconciliation reconciliation, CrdOperator<KubernetesClient, KafkaConnector, KafkaConnectorList> resourceOperator,
                                                            CustomResource resource) {
         return resourceOperator
             .getAsync(resource.getMetadata().getNamespace(), resource.getMetadata().getName())
@@ -964,7 +715,7 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
      *
      * @return  Future which completes then the status is updated
      */
-    protected <U extends CustomResource<?, V>, V extends Status, W extends DefaultKubernetesResourceList<U>> Future<Void>
+    protected static <U extends CustomResource<?, V>, V extends Status, W extends DefaultKubernetesResourceList<U>> Future<Void>
         maybeUpdateStatusCommon(CrdOperator<KubernetesClient, U, W> resourceOperator,
                                 U resource,
                                 Reconciliation reconciliation,
@@ -1018,17 +769,19 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
     }
 
     /**
-     * Creates the ClusterRoleBinding required for the init container used for client rack-awareness.
-     * The init-container needs to be able to read the labels from the node it is running on to be able
-     * to determine the `client.rack` option.
-     *
-     * @param reconciliation The reconciliation
-     * @param crbName ClusterRoleBinding name
-     * @param crb ClusterRoleBinding
-     *
-     * @return Future for tracking the asynchronous result of the ClusterRoleBinding reconciliation
+     * Utility class used to transfer data. It is not immutable, so record cannot be used.
      */
-    protected Future<ReconcileResult<ClusterRoleBinding>> connectInitClusterRoleBinding(Reconciliation reconciliation, String crbName, ClusterRoleBinding crb) {
-        return ReconcilerUtils.withIgnoreRbacError(reconciliation, clusterRoleBindingOperations.reconcile(reconciliation, crbName, crb), crb);
+    static protected class ConnectorStatusAndConditions {
+        Map<String, Object> statusResult;
+        List<String> topics;
+        List<Condition> conditions;
+        AutoRestartStatus autoRestart;
+
+        ConnectorStatusAndConditions(Map<String, Object> statusResult, List<String> topics, List<Condition> conditions, AutoRestartStatus autoRestart) {
+            this.statusResult = statusResult;
+            this.topics = topics;
+            this.conditions = conditions;
+            this.autoRestart = autoRestart;
+        }
     }
 }
